+++ docs/release_notes.rst
@@ -8,9 +8,14 @@ Release notes for **Shopit**.
 0.1.4
 =====
 
+* Add `description` field to categorization models.
 * Move variant generator methods from admin to the model. Now ``create_all_variants`` and ``create_variant`` are
   available on the model.
 * Update add to cart ``get_context`` to ensure correct product translation is returned.
+
+.. attention::
+
+    Requires ``python manage.py migrate shopit`` to create description field on categorization models.
 
 0.1.3
 =====